Cleaned up Retention offers feature flag
closes https://linear.app/ghost/issue/BER-3387
closes https://linear.app/ghost/issue/BER-3370

- Retention offers have been pushed to GA in commit e66644b39b6785a27e5494adc9640d22f34e68b3
- This change cleans up the feature flag and related dead code
