+++ apps/admin-x-settings/src/components/settings/growth/offers/offers-container-modal.tsx
@@ -3,9 +3,7 @@ import EditOfferModal from './edit-offer-modal';
 import EditRetentionOfferModal from './edit-retention-offer-modal';
 import NiceModal from '@ebay/nice-modal-react';
 import OfferSuccess from './offer-success';
-import useFeatureFlag from '../../../../hooks/use-feature-flag';
 import {OffersIndexModal} from './offers-index';
-import {OffersIndexModal as OffersIndexRetentionModal} from './offers-index-retention';
 import {useRouting} from '@tryghost/admin-x-framework/routing';
 
 type OffersRouteHandlerProps = {
@@ -13,14 +11,12 @@ type OffersRouteHandlerProps = {
 };
 
 const OffersRouteHandler: React.FC<OffersRouteHandlerProps> = ({route}) => {
-    const retentionOffersEnabled = useFeatureFlag('retentionOffers');
-
     if (route === 'offers/new') {
         return <AddOfferModal />;
-    } else if (retentionOffersEnabled && route.startsWith('offers/edit/retention/') && route.length > 'offers/edit/retention/'.length) {
+    } else if (route.startsWith('offers/edit/retention/') && route.length > 'offers/edit/retention/'.length) {
         const retentionId = route.split('/').pop();
         return <EditRetentionOfferModal id={retentionId ? retentionId : ''} />;
-    } else if (!retentionOffersEnabled && route.startsWith('offers/edit/retention')) {
+    } else if (route === 'offers/edit/retention/' || route === 'offers/edit/retention') {
         return <OffersIndexModal />;
     } else if (route.startsWith('offers/edit/') && route.length > 'offers/edit/'.length) {
         const offerId = route.split('/').pop();
@@ -28,8 +24,8 @@ const OffersRouteHandler: React.FC<OffersRouteHandlerProps> = ({route}) => {
     } else if (route.startsWith('offers/success/') && route.length > 'offers/success/'.length) {
         const offerId = route.split('/').pop();
         return <OfferSuccess id={offerId ? offerId : ''} />;
-    } else if (route === 'offers/edit') {
-        return retentionOffersEnabled ? <OffersIndexRetentionModal /> : <OffersIndexModal />;
+    } else if (route === 'offers/edit' || route === 'offers/edit/') {
+        return <OffersIndexModal />;
     }
 };
