+++ apps/admin-x-settings/src/components/settings/growth/offers/offers-index.tsx
@@ -1,27 +1,21 @@
-import {Button, type Tab, TabView} from '@tryghost/admin-x-design-system';
-import {ButtonGroup, type ButtonProps, showToast} from '@tryghost/admin-x-design-system';
+import {Button, type ButtonProps, showToast} from '@tryghost/admin-x-design-system';
+import {ButtonGroup} from '@tryghost/admin-x-design-system';
 import {Icon} from '@tryghost/admin-x-design-system';
+import {LucideIcon} from '@tryghost/shade';
 import {Modal} from '@tryghost/admin-x-design-system';
-import {SortMenu} from '@tryghost/admin-x-design-system';
+import {Popover} from '@tryghost/admin-x-design-system';
+import {type RetentionOffer, getRetentionOffers} from './offers-retention';
 import {type Tier, getPaidActiveTiers, useBrowseTiers} from '@tryghost/admin-x-framework/api/tiers';
-import {Tooltip} from '@tryghost/admin-x-design-system';
+import {Toggle} from '@tryghost/admin-x-design-system';
+import {createOfferRedemptionFilterUrl, createOfferRedemptionsFilterUrl} from './offer-helpers';
 import {currencyToDecimal, getSymbol} from '../../../../utils/currency';
-import {getHomepageUrl} from '@tryghost/admin-x-framework/api/site';
 import {numberWithCommas} from '../../../../utils/helpers';
 import {useBrowseOffers} from '@tryghost/admin-x-framework/api/offers';
-import {useGlobalData} from '../../../providers/global-data-provider';
 import {useModal} from '@ebay/nice-modal-react';
+import {useOffersShowArchived, useSortingState} from '../../../providers/settings-app-provider';
 import {useRouting} from '@tryghost/admin-x-framework/routing';
-import {useSortingState} from '../../../providers/settings-app-provider';
-import {useState} from 'react';
 
 export type OfferType = 'percent' | 'fixed' | 'trial';
-
-export const createRedemptionFilterUrl = (id: string): string => {
-    const baseHref = '/ghost/#/members';
-    const filterValue = `offer_redemptions:[${id}]`;
-    return `${baseHref}?filter=${encodeURIComponent(filterValue)}`;
-};
 
 export const getOfferCadence = (cadence: string): string => {
     return cadence === 'month' ? 'monthly' : 'yearly';
@@ -75,62 +69,167 @@ export const getOfferDiscount = (type: string, amount: number, cadence: string, 
     };
 };
 
-export const CopyLinkButton: React.FC<{offerCode: string}> = ({offerCode}) => {
-    const [isCopied, setIsCopied] = useState(false);
-    const {siteData} = useGlobalData();
-
-    const handleCopyClick = (e?: React.MouseEvent<HTMLElement, MouseEvent>) => {
-        e?.stopPropagation();
-        const offerLink = `${getHomepageUrl(siteData!)}${offerCode}`;
-        navigator.clipboard.writeText(offerLink);
-        setIsCopied(true);
-        setTimeout(() => setIsCopied(false), 2000);
-    };
-
-    return <Tooltip containerClassName='group-hover:opacity-100 opacity-0 inline-flex items-center -mr-1 justify-center leading-none w-5 h-5' content={isCopied ? 'Copied' : 'Copy link'} size='sm'><Button color='clear' hideLabel={true} icon={isCopied ? 'check-circle' : 'hyperlink-circle'} iconColorClass={isCopied ? 'text-green w-[14px] h-[14px]' : 'w-[14px] h-[14px]'} label={isCopied ? 'Copied' : 'Copy'} unstyled={true} onClick={handleCopyClick} /></Tooltip>;
+const OffersFilterPopover: React.FC<{
+    showArchived: boolean;
+    setShowArchived: (show: boolean) => void;
+    sortOption: string;
+    sortDirection: string;
+    onSortChange: (option: string) => void;
+    onDirectionChange: () => void;
+}> = ({showArchived, setShowArchived, sortOption, sortDirection, onSortChange, onDirectionChange}) => {
+    return (
+        <Popover
+            position='end'
+            trigger={
+                <Button aria-label="Filter options" className='p-1 hover:text-black dark:hover:text-white' label={<LucideIcon.ListFilter size={16} strokeWidth={1.5} />} unstyled={true} />
+            }
+        >
+            <div className='flex min-w-[200px] flex-col p-1 normal-case'>
+                <div className='cursor-default select-none pl-3 pt-2 text-xs font-medium uppercase tracking-wide text-grey-700'>Sort by</div>
+                <div className='flex flex-col py-1'>
+                    {[
+                        {id: 'date-added', label: 'Date added'},
+                        {id: 'name', label: 'Name'},
+                        {id: 'redemptions', label: 'Redemptions'}
+                    ].map(item => (
+                        <div
+                            key={item.id}
+                            className='group relative mx-1 flex items-center rounded-[2.5px] hover:bg-grey-100 dark:hover:bg-grey-800'
+                        >
+                            <button
+                                className='flex w-full cursor-pointer items-center px-8 py-1.5 pr-12 text-left text-sm text-black dark:text-white'
+                                type='button'
+                                onClick={() => onSortChange(item.id)}
+                            >
+                                {sortOption === item.id && <Icon className='absolute left-2' name='check' size='xs' />}
+                                {item.label}
+                            </button>
+                            {sortOption === item.id && (
+                                <button
+                                    className='absolute right-1 flex size-6 cursor-pointer items-center justify-center rounded-full hover:bg-grey-300 dark:hover:bg-grey-700'
+                                    title={sortDirection === 'asc' ? 'Ascending' : 'Descending'}
+                                    type='button'
+                                    onClick={() => onDirectionChange()}
+                                >
+                                    <Icon name={sortDirection === 'asc' ? 'arrow-up' : 'arrow-down'} size='xs' />
+                                </button>
+                            )}
+                        </div>
+                    ))}
+                </div>
+                <div className='-mx-1 mt-1 border-t border-t-grey-200 dark:border-t-grey-800'>
+                    <div className='group relative mx-2 mt-1 flex items-center rounded-[2.5px] py-1'>
+                        <div className='flex w-full items-center px-8 py-1.5 pr-2 text-sm text-black dark:text-white'>
+                            <LucideIcon.Archive className='absolute left-2 -mt-0.5 text-black dark:text-white' size={14} strokeWidth={1.5} />
+                            <div className='grow [&>div]:w-full'>
+                                <Toggle
+                                    checked={showArchived}
+                                    direction='rtl'
+                                    label='Show archived'
+                                    labelClasses='text-sm text-black dark:text-white'
+                                    onChange={(e) => {
+                                        setShowArchived(e.target.checked);
+                                    }}
+                                />
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </Popover>
+    );
 };
 
-export const EmptyState: React.FC<{title?: string, description: string, buttonAction: () => void, buttonLabel: string}> = ({title = 'No offers found', description, buttonAction, buttonLabel}) => (
-    <div className='flex h-full grow flex-col items-center justify-center text-center'>
-        <Icon className='-mt-14' colorClass='text-grey-700 -mt-6' name='tags-block' size='xl' />
-        <h1 className='mt-6 text-2xl'>{title}</h1>
-        <p className='mt-3 max-w-[420px] text-[1.6rem]'>{description}</p>
-        <Button className="mt-8" color="grey" label={buttonLabel} onClick={buttonAction}></Button>
-    </div>
-);
+const RetentionOfferRow: React.FC<{
+    offer: RetentionOffer;
+    onClick: () => void;
+}> = ({offer, onClick}) => {
+    const redemptionFilterUrl = offer.redemptions > 0 && offer.redemptionOfferIds.length > 0
+        ? createOfferRedemptionsFilterUrl(offer.redemptionOfferIds)
+        : undefined;
 
-export const OffersIndexModal = () => {
+    return (
+        <tr className='group relative border-b border-b-grey-200 dark:border-grey-800' data-testid='retention-offer-item'>
+            <td className='sticky left-0 z-10 bg-white p-0 dark:bg-black'>
+                <button className='block w-full cursor-pointer p-5 pl-0 text-left' type="button" onClick={onClick}>
+                    <span className='font-semibold'>{offer.name}</span><br />
+                    <span className='text-sm text-grey-700'>{offer.description}</span>
+                </button>
+            </td>
+            <td className='whitespace-nowrap p-0 text-sm'>
+                <button className='block w-full cursor-pointer p-5 text-left' type="button" onClick={onClick}>
+                    {offer.terms ? (
+                        <>
+                            <span className='text-[1.3rem] font-medium uppercase'>{offer.terms}</span><br />
+                            <span className='text-grey-700'>{offer.termsDetail}</span>
+                        </>
+                    ) : (
+                        <span className='text-grey-700'>&ndash;</span>
+                    )}
+                </button>
+            </td>
+            <td className='whitespace-nowrap p-0 text-sm'>
+                <button className='block w-full cursor-pointer p-5 text-left' type="button" onClick={onClick}>
+                    <span className='text-grey-700'>&ndash;</span>
+                </button>
+            </td>
+            <td className='whitespace-nowrap p-0 text-sm'>
+                {redemptionFilterUrl ? (
+                    <a
+                        className='block cursor-pointer p-5 hover:underline'
+                        data-testid={`retention-redemptions-link-${offer.id}`}
+                        href={redemptionFilterUrl}
+                    >
+                        {offer.redemptions}
+                    </a>
+                ) : (
+                    <button
+                        className='block w-full cursor-pointer p-5 text-left'
+                        data-testid={`retention-redemptions-link-${offer.id}`}
+                        type="button"
+                        onClick={onClick}
+                    >
+                        {offer.redemptions}
+                    </button>
+                )}
+            </td>
+            <td className='whitespace-nowrap p-0 text-sm'>
+                <button className='block w-full cursor-pointer p-5 text-left' type="button" onClick={onClick}>
+                    {offer.status === 'active' ? (
+                        <span className='inline-flex items-center rounded-full bg-[rgba(48,207,67,0.15)] px-2 py-0.5 text-2xs font-semibold uppercase tracking-wide text-green'>Active</span>
+                    ) : (
+                        <span className='inline-flex items-center rounded-full bg-grey-200 px-2 py-0.5 text-2xs font-semibold uppercase tracking-wide text-grey-700 dark:bg-grey-900 dark:text-grey-500'>Inactive</span>
+                    )}
+                </button>
+            </td>
+        </tr>
+    );
+};
+
+export const OffersIndexModal: React.FC = () => {
     const modal = useModal();
     const {updateRoute} = useRouting();
-    const {data: {offers: allOffers = []} = {}, isFetching: isFetchingOffers} = useBrowseOffers();
+    const {data: {offers: allOffers = []} = {}} = useBrowseOffers();
     const {data: {tiers: allTiers} = {}} = useBrowseTiers();
     const signupOffers = allOffers.filter(offer => offer.redemption_type === 'signup');
-    const activeOffers = signupOffers.filter((offer) => {
-        const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
-        return (offer.status === 'active' && offerTier && offerTier.active === true);
-    });
-    const archivedOffers = signupOffers.filter((offer) => {
-        const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
-        return (offer.status === 'archived' || (offerTier && offerTier.active === false));
-    });
-
-    let offersTabs: Tab[] = [
-        {id: 'active', title: 'Active'},
-        {id: 'archived', title: 'Archived'}
-    ];
-
+    const retentionOffers = getRetentionOffers(allOffers);
     const {sortingState, setSortingState} = useSortingState();
     const offersSorting = sortingState?.find(sorting => sorting.type === 'offers');
 
-    const [selectedTab, setSelectedTab] = useState('active');
+    const {offersShowArchived: showArchived, setOffersShowArchived: setShowArchived} = useOffersShowArchived();
 
     const sortOption = offersSorting?.option || 'date-added';
     const sortDirection = offersSorting?.direction || 'desc';
 
-    const handleOfferEdit = (id:string) => {
-        // TODO: implement
-        sessionStorage.setItem('editOfferPageSource', 'offersIndex');
+    const handleOfferEdit = (id: string) => {
+        if (!id) {
+            return;
+        }
         updateRoute(`offers/edit/${id}`);
+    };
+
+    const handleRetentionOfferEdit = (id: string) => {
+        updateRoute(`offers/edit/retention/${id}`);
     };
 
     const sortedOffers = signupOffers
@@ -148,49 +247,41 @@ export const OffersIndexModal = () => {
 
     const paidActiveTiers = getPaidActiveTiers(allTiers || []);
 
-    const listLayoutOutput = <div className='overflow-x-auto'>
-        <table className='m-0 w-full'>
-            {(selectedTab === 'active' && activeOffers.length > 0) || (selectedTab === 'archived' && archivedOffers.length > 0) ?
-                <tr className='border-b border-b-grey-300 dark:border-grey-800'>
-                    <th className='px-5 py-2.5 pl-0 text-xs font-normal text-grey-700'>{selectedTab === 'active' ? activeOffers.length : archivedOffers.length} {selectedTab === 'active' ? (activeOffers.length !== 1 ? 'offers' : 'offer') : (archivedOffers.length !== 1 ? 'offers' : 'offer')}</th>
-                    <th className='px-5 py-2.5 text-xs font-normal text-grey-700'>Terms</th>
-                    <th className='px-5 py-2.5 text-xs font-normal text-grey-700'>Price</th>
-                    <th className='px-5 py-2.5 text-xs font-normal text-grey-700'>Redemptions</th>
-                    <th className='min-w-[80px] px-5 py-2.5 pr-0 text-xs font-normal text-grey-700'></th>
-                </tr> :
-                null
-            }
-            {sortedOffers.filter((offer) => {
-                const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
-                return (selectedTab === 'active' && (offer.status === 'active' && offerTier && offerTier.active === true)) ||
-                (selectedTab === 'archived' && (offer.status === 'archived' || (offerTier && offerTier.active === false)));
-            }).map((offer) => {
-                const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
+    const filteredSignupOffers = sortedOffers.filter((offer) => {
+        const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
+        const isActive = offer.status === 'active' && offerTier && offerTier.active === true;
+        const isArchived = offer.status === 'archived' || (offerTier && offerTier.active === false);
 
-                if (!offerTier) {
-                    return null;
-                }
+        if (isActive) {
+            return true;
+        }
+        if (showArchived && isArchived) {
+            return true;
+        }
+        return false;
+    });
 
-                const isTierArchived = offerTier?.active === false;
+    const handleSortChange = (selectedOption: string) => {
+        setSortingState?.([{
+            type: 'offers',
+            option: selectedOption,
+            direction: sortDirection
+        }]);
+    };
 
-                const {discountOffer, originalPriceWithCurrency, updatedPriceWithCurrency} = getOfferDiscount(offer.type, offer.amount, offer.cadence, offer.currency || 'USD', offerTier);
+    const handleDirectionChange = () => {
+        const newDirection = sortDirection === 'asc' ? 'desc' : 'asc';
+        setSortingState?.([{
+            type: 'offers',
+            option: sortOption,
+            direction: newDirection
+        }]);
+    };
 
-                return (
-                    <tr className={`group relative scale-100 border-b border-b-grey-200 dark:border-grey-800`} data-testid="offer-item">
-                        <td className={`${isTierArchived ? 'opacity-50' : ''} p-0`}><a className={`block ${isTierArchived ? 'cursor-default select-none' : 'cursor-pointer'} p-5 pl-0`} onClick={!isTierArchived ? () => handleOfferEdit(offer?.id ? offer.id : '') : () => {}}><span className='font-semibold'>{offer?.name}</span><br /><span className='text-sm text-grey-700'>{offerTier.name} {getOfferCadence(offer.cadence)}</span></a></td>
-                        <td className={`${isTierArchived ? 'opacity-50' : ''} whitespace-nowrap p-0 text-sm`}><a className={`block ${isTierArchived ? 'cursor-default select-none' : 'cursor-pointer'} p-5`} onClick={!isTierArchived ? () => handleOfferEdit(offer?.id ? offer.id : '') : () => {}}><span className='text-[1.3rem] font-medium uppercase'>{discountOffer}</span><br /><span className='text-grey-700'>{offer.type !== 'trial' ? getOfferDuration(offer.duration) : 'Trial period'}</span></a></td>
-                        <td className={`${isTierArchived ? 'opacity-50' : ''} whitespace-nowrap p-0 text-sm`}><a className={`block ${isTierArchived ? 'cursor-default select-none' : 'cursor-pointer'} p-5`} onClick={!isTierArchived ? () => handleOfferEdit(offer?.id ? offer.id : '') : () => {}}><span className='font-medium'>{updatedPriceWithCurrency}</span> {offer.type !== 'trial' ? <span className='relative text-xs text-grey-700 before:absolute before:-inset-x-0.5 before:top-1/2 before:rotate-[-20deg] before:border-t before:content-[""]'>{originalPriceWithCurrency}</span> : null}</a></td>
-                        <td className={`${isTierArchived ? 'opacity-50' : ''} w-[120px] whitespace-nowrap p-0 text-sm`}><a className={`block ${isTierArchived ? 'cursor-default select-none' : 'cursor-pointer'} p-5 ${offer.redemption_count === 0 ? '' : 'hover:underline'}`} href={offer.redemption_count > 0 ? createRedemptionFilterUrl(offer.id ? offer.id : '') : undefined} onClick={offer.redemption_count === 0 ? !isTierArchived ? () => handleOfferEdit(offer?.id ? offer.id : '') : () => {} : () => {}}>{offer.redemption_count}</a></td>
-                        <td className={`${isTierArchived ? 'opacity-50' : ''} w-[120px] whitespace-nowrap p-5 pr-8 text-right text-sm leading-none`}>{!isTierArchived ? <CopyLinkButton offerCode={offer.code} /> : null}</td>
-                        {isTierArchived ?
-                            <div className='absolute right-0 top-[11px] whitespace-nowrap rounded-sm bg-black px-2 py-0.5 text-xs leading-normal text-white opacity-0 transition-all group-hover:opacity-100 dark:bg-grey-950'>This offer is disabled, because <br /> it is tied to an archived tier.</div> :
-                            null
-                        }
-                    </tr>
-                );
-            })}
-        </table>
-    </div>;
+    const isOfferArchived = (offer: typeof signupOffers[0]) => {
+        const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
+        return offer.status === 'archived' || (offerTier && offerTier.active === false);
+    };
 
     const buttons: ButtonProps[] = [
         {
@@ -205,7 +296,7 @@ export const OffersIndexModal = () => {
             key: 'new-offer',
             icon: 'add',
             label: 'New offer',
-            color: 'green',
+            color: 'green' as const,
             onClick: () => {
                 if (paidActiveTiers.length === 0) {
                     showToast({
@@ -218,6 +309,77 @@ export const OffersIndexModal = () => {
             }
         }
     ];
+
+    const listLayoutOutput = <div className='overflow-x-auto'>
+        <table className='m-0 min-w-[900px]'>
+            <colgroup>
+                <col className='w-[25%]' />
+                <col className='w-[200px]' />
+                <col className='w-[200px]' />
+                <col className='w-[200px]' />
+                <col className='w-[220px]' />
+            </colgroup>
+            <thead>
+                <tr className='border-b border-b-grey-200 dark:border-grey-800'>
+                    <th className='sticky left-0 z-10 bg-white p-0 pb-2.5 text-left text-xs font-medium uppercase tracking-wide text-grey-700 dark:bg-black'>Name</th>
+                    <th className='p-0 pb-2.5 pl-5 text-left text-xs font-medium uppercase tracking-wide text-grey-700'>Terms</th>
+                    <th className='p-0 pb-2.5 pl-5 text-left text-xs font-medium uppercase tracking-wide text-grey-700'>Price</th>
+                    <th className='p-0 pb-2.5 pl-5 text-left text-xs font-medium uppercase tracking-wide text-grey-700'>Redemptions</th>
+                    <th className='p-0 pb-2.5 pl-5 text-left text-xs font-medium uppercase tracking-wide text-grey-700'>
+                        <span className='flex items-center justify-between'>
+                            Status
+                            <OffersFilterPopover
+                                setShowArchived={setShowArchived}
+                                showArchived={showArchived}
+                                sortDirection={sortDirection}
+                                sortOption={sortOption}
+                                onDirectionChange={handleDirectionChange}
+                                onSortChange={handleSortChange}
+                            />
+                        </span>
+                    </th>
+                </tr>
+            </thead>
+            <tbody>
+                {retentionOffers.map(offer => (
+                    <RetentionOfferRow
+                        key={offer.id}
+                        offer={offer}
+                        onClick={() => handleRetentionOfferEdit(offer.id)}
+                    />
+                ))}
+                {filteredSignupOffers.map((offer) => {
+                    const offerTier = allTiers?.find(tier => tier.id === offer?.tier?.id);
+
+                    if (!offerTier) {
+                        return null;
+                    }
+
+                    const archived = isOfferArchived(offer);
+
+                    const {discountOffer, originalPriceWithCurrency, updatedPriceWithCurrency} = getOfferDiscount(offer.type, offer.amount, offer.cadence, offer.currency || 'USD', offerTier);
+
+                    return (
+                        <tr key={offer.id} className={`group relative border-b border-b-grey-200 dark:border-grey-800 ${archived ? 'opacity-60' : ''}`} data-testid="offer-item">
+                            <td className='sticky left-0 z-10 bg-white p-0 dark:bg-black'><a className='block cursor-pointer p-5 pl-0' onClick={() => handleOfferEdit(offer.id)}><span className='font-semibold'>{offer?.name}</span><br /><span className='text-sm text-grey-700'>{offerTier.name} {getOfferCadence(offer.cadence)}</span></a></td>
+                            <td className='whitespace-nowrap p-0 text-sm'><a className='block cursor-pointer p-5' onClick={() => handleOfferEdit(offer.id)}><span className='text-[1.3rem] font-medium uppercase'>{discountOffer}</span><br /><span className='text-grey-700'>{offer.type !== 'trial' ? getOfferDuration(offer.duration) : 'Trial period'}</span></a></td>
+                            <td className='whitespace-nowrap p-0 text-sm'><a className='block cursor-pointer p-5' onClick={() => handleOfferEdit(offer.id)}><span className='font-medium'>{updatedPriceWithCurrency}</span> {offer.type !== 'trial' ? <span className='relative text-xs text-grey-700 before:absolute before:-inset-x-0.5 before:top-1/2 before:rotate-[-20deg] before:border-t before:content-[""]'>{originalPriceWithCurrency}</span> : null}</a></td>
+                            <td className='whitespace-nowrap p-0 text-sm'><a className={`block cursor-pointer p-5 ${offer.redemption_count === 0 ? '' : 'hover:underline'}`} href={offer.redemption_count > 0 && offer.id ? createOfferRedemptionFilterUrl(offer.id) : undefined} onClick={offer.redemption_count === 0 && offer.id ? () => handleOfferEdit(offer.id) : undefined}>{offer.redemption_count}</a></td>
+                            <td className='whitespace-nowrap p-0 text-sm'>
+                                <a className='block cursor-pointer p-5' onClick={() => handleOfferEdit(offer.id)}>
+                                    {archived ? (
+                                        <span className='inline-flex items-center rounded-full bg-grey-200 px-2 py-0.5 text-2xs font-semibold uppercase tracking-wide text-grey-700 dark:bg-grey-900 dark:text-grey-500'>Archived</span>
+                                    ) : (
+                                        <span className='inline-flex items-center rounded-full bg-[rgba(48,207,67,0.15)] px-2 py-0.5 text-2xs font-semibold uppercase tracking-wide text-green'>Active</span>
+                                    )}
+                                </a>
+                            </td>
+                        </tr>
+                    );
+                })}
+            </tbody>
+        </table>
+    </div>;
 
     return <Modal
         afterClose={() => {
@@ -235,62 +397,6 @@ export const OffersIndexModal = () => {
         width={1140}
     >
         <div className='flex h-full flex-col pt-8'>
-            <header>
-                <TabView
-                    selectedTab={selectedTab}
-                    tabs={offersTabs}
-                    topRightContent={
-                        (selectedTab === 'active' && activeOffers.length > 0) || (selectedTab === 'archived' && archivedOffers.length > 0) ?
-                            <div className='pt-1'>
-                                <SortMenu
-                                    direction={sortDirection as 'asc' | 'desc'}
-                                    items={[
-                                        {id: 'date-added', label: 'Date added', selected: sortOption === 'date-added', direction: sortDirection as 'asc' | 'desc'},
-                                        {id: 'name', label: 'Name', selected: sortOption === 'name', direction: sortDirection as 'asc' | 'desc'},
-                                        {id: 'redemptions', label: 'Redemptions', selected: sortOption === 'redemptions', direction: sortDirection as 'asc' | 'desc'}
-                                    ]}
-                                    position='end'
-                                    triggerButtonProps={{
-                                        link: true
-                                    }}
-                                    onDirectionChange={(selectedDirection) => {
-                                        const newDirection = selectedDirection === 'asc' ? 'desc' : 'asc';
-                                        setSortingState?.([{
-                                            type: 'offers',
-                                            option: sortOption,
-                                            direction: newDirection
-                                        }]);
-                                    }}
-                                    onSortChange={(selectedOption) => {
-                                        setSortingState?.([{
-                                            type: 'offers',
-                                            option: selectedOption,
-                                            direction: sortDirection
-                                        }]);
-                                    }}
-                                />
-                            </div> :
-                            null
-                    }
-                    onTabChange={setSelectedTab}
-                />
-            </header>
-            {selectedTab === 'active' && activeOffers.length === 0 && !isFetchingOffers ?
-                <EmptyState
-                    buttonAction={() => updateRoute('offers/new')}
-                    buttonLabel='Create an offer'
-                    description='Grow your audience with discounts or free trials.'
-                /> :
-                null
-            }
-            {selectedTab === 'archived' && archivedOffers.length === 0 && !isFetchingOffers ?
-                <EmptyState
-                    buttonAction={() => setSelectedTab('active')}
-                    buttonLabel='Back to active'
-                    description='All archived offers will be shown here.'
-                /> :
-                null
-            }
             {listLayoutOutput}
         </div>
     </Modal>;
