+++ ghost/core/core/shared/labs.js
@@ -25,8 +25,7 @@ const GA_FEATURES = [
     'explore',
     'commentModeration',
     'featurebaseFeedback',
-    'welcomeEmailEditor',
-    'retentionOffers'
+    'welcomeEmailEditor'
 ];
 
 // These features are considered publicly available and can be enabled/disabled by users
